update function deleteCellAndAllChilds, and construction of unit tests of cell.methods in progress.

+++ ecosystem_mindmap/backend/database/cells/cells.methods.ts
@@ -1,95 +1,15 @@
 // interface typing.
 import { INewCell, IGetCells, ICell, IGetCellByPropsId,IgetEcoSystemByStemCellId,
-        IUpdatePropsCellResp, INewParentsTreeOfTheCellResp, 
-        IGetAllIdOfChildCellsResp,
+        IUpdatePropsCellResp,
+        ICellSchema
      } from './cells.types';
 
 // Local file.
 import { CellModel, ParentsTreeOfTheCellModel } from './cells.model';
-
-/* ------------------------------------------------------------
------------------- helpers functions --------------------------
------------------------------------------------------------- */
-
-/*
-    Create new cell in database.
-    Model : CellModel.
-unit test _ done. */
-export async function newCell ( 
-    title: string, description: string, position: number,
-    idStemCell: string,stemCell: boolean ): Promise<INewCell> {
-
-    const requestType = 'Create new Cell.';
-    
-    const newCell = new CellModel({
-        title, description, position, stemCell, 
-        idStemCell
-    });
-    
-    const newCellCreated:INewCell = await newCell.save()
-        .then(newCellCreated => {return {
-                request_type: requestType,
-                error: false, 
-                cell_created: newCellCreated
-            }})
-        .catch(error => {return {
-                request_type: requestType,
-                error: true, 
-                message: error
-            }})
-    return newCellCreated;
-};
-
-/* 
-    Create parent tree of the cell in database.
-    Model : ParentsTreeOfTheCellModel.
-unit test _ done. */
-export async function newParentsTreeOfTheCell (parentsArray: string[], cell_id: string): Promise<INewParentsTreeOfTheCellResp> {
-    
-    const requestType = 'create new parents tree of the cell.'; 
-    
-    const newParentsTreeOfTheCell = new ParentsTreeOfTheCellModel({
-        cellId: cell_id,
-        cellLevel: parentsArray.length,
-        parentsIdList: parentsArray,
-    });
-    
-    const parentsTreeOfTheCellCreated = await newParentsTreeOfTheCell.save()
-            .then(newParentsTreeCreated => {return {
-                    request_type: requestType,
-                    error: false,
-                    parents_tree: newParentsTreeCreated,
-                }})
-            .catch(error => {return {
-                    request_type: requestType,
-                    error: true,
-                    message: error,
-                }});
-    return parentsTreeOfTheCellCreated;
-};
-
-/* 
-    Get the cell(s) by idStemCell property specified as a parameter of this function.
-    { idStemCell: id} 
-    Model : CellModel.
-unit test _ done. */
-export async function getCellsByPropsIdStemCell (id: string): Promise<IGetCells> {
-    
-    const requestType = 'Get cell(s) by idStemCell.';
-    
-    const cells_Request = await CellModel.find({idStemCell: id})
-        .then(cells_Response => {return  {
-                request_type: requestType,
-                error: false, 
-                cells_Request: cells_Response,
-            }})
-        .catch(error => {return {
-                request_type: requestType,
-                error: true, 
-                message: error
-            }})
-    return cells_Request;
-};
+import { newCell, getCellsByPropsIdStemCell, newParentsTreeOfTheCell, 
+        getAllIdOfChildCells, deleteAllChildrenCellsOfTheCellDeleted,
+        deleteAllParentsTreesOfTheCellDeleted
+} from './cells.helpers.methods'
 
 /* 
     Get the cell with the _id property specified as a parameter of this function.
@@ -113,90 +33,7 @@ export async function getCellByProps_Id (id: string): Promise<IGetCellByPropsId>
                 message: error
             }})
     return cells_Request;
-}
-
-/* 
-    Get all document which contains the id in the parensIdList property.
-    This request is used to generate the list of each document containing
-    the id, the level of the cell id passed as a parameter.
-    {parentsIdList: id}
-    Model : ParentsTreeOfTheCellModel.
-unit test _ done. */
-export async function getAllIdOfChildCells (id: string): Promise<IGetAllIdOfChildCellsResp> {
-    
-    const requestType = 'get all id of childs Cells.'
-   
-    const childsIdList = await ParentsTreeOfTheCellModel.find({parentsIdList: id})
-            .then(getChildsCells => {return {
-                    request_type: requestType,
-                    error: false,
-                    parents_tree: getChildsCells,
-                }})
-            .catch(error => {return {
-                    request_type: requestType,
-                    error: true,
-                    message: error,
-                }});
-    return childsIdList;
 };
-
-/* 
-    Delete all children cells of the cell deleted.
-    Model : CellModel.
-unit test _ done. */
-export async function deleteAllChildrenCellsOfTheCellDeleted (chidrenIdList: IGetAllIdOfChildCellsResp) {
-    
-    chidrenIdList.parents_tree.map(async currentId => {
-        await CellModel.findByIdAndDelete(currentId.cellId)
-            .then(deleteChildCell => {
-                return {
-                    request_type: `Deleted child cell ${currentId.cellId} of model CellModel.`,
-                    error: false,
-                    child_cell_deleted: deleteChildCell,
-                }
-            })
-            .catch(error => {
-                return {
-                    request_type: `Deleted child cell ${currentId.cellId} of model CellModel.`,
-                    error: true,
-                    message: error,
-                }
-            });
-    });
-};
-
-/* 
-    Delete all parents trees of the children cells of the cell deleted.
-    Model : ParentsTreeOfTheCellModel.
-unit test _ done. */
-export async function deleteAllParentsTreesOfTheCellDeleted (childrenIdList: IGetAllIdOfChildCellsResp) {
-    childrenIdList.parents_tree.map(async currentId => {
-        await ParentsTreeOfTheCellModel.findByIdAndDelete(currentId._id)
-            .then(deleteParentTree => {
-                return {
-                    request_type: `Delete parent tree ${currentId._id} of model ParentsTreeOfTheCellModel.`,
-                    error: false,
-                    parent_tree_deleted: deleteParentTree,
-                }
-            })
-            .catch(error => {
-                return {
-                    request_type: `Delete parent tree ${currentId._id} of model ParentsTreeOfTheCellModel.`,
-                    error: true,
-                    message: error,
-                }
-            });
-    });
-};
-
-/* __________________________________________________________ */
-
-
-/* ------------------------------------------------------------------------------------------------------
----------------------------------------------------------------------------------------------------------
------------------------- function used in the socket.io communication -----------------------------------
----------------------------------------------------------------------------------------------------------
---------------------------------------------------------------------------------------------------------- */
 
 /* 
     Get ecosystem, to define stem cell & cell(s) to display.
@@ -252,7 +89,7 @@ export async function getEcoSystemByStemCellId (stemCell_id: string, parentIsMin
 /* 
     Create default stem cell. 
     If the mind map is empty this function is call.
- */
+unit test _ done. */
 export async function createDefaultStemCell (stemCellId: string): Promise<INewCell> {
     
     const createDefaultStemCell: INewCell = await newCell(
@@ -269,8 +106,8 @@ export async function createDefaultStemCell (stemCellId: string): Promise<INewCe
 /* 
     Add new cell and update the position of the cell(s) with the same stem cell.
     Add parent tree of the cell added.
- */
-export async function addCell (cell: ICell, parentTree: string[]): Promise<INewCell> {
+unit test _ done. */
+export async function addCell (cell: ICellSchema, parentTree: string[]): Promise<INewCell> {
     
     const cells: IGetCells = await getCellsByPropsIdStemCell(cell.idStemCell);
     
@@ -304,7 +141,7 @@ export async function addCell (cell: ICell, parentTree: string[]): Promise<INewC
 /* 
     Update properties of the cell.
     If title or description is changed, this function is call.
- */
+unit test _ done. */
 export async function updatePropsCellById (cellUpdated:ICell):  Promise<IUpdatePropsCellResp>  {
     
     const requestType = 'update props cell';
@@ -326,34 +163,43 @@ export async function updatePropsCellById (cellUpdated:ICell):  Promise<IUpdateP
     Delete cell and delete all children of that cell.
     The position of the cell(s) with the same stem cell are updated.
  */
-export async function deleteCellAndAllChilds (cell_id: string, stemCell_id: string) {
-    
-    const cells: IGetCells = await getCellsByPropsIdStemCell(stemCell_id);
+export async function deleteCellAndAllChilds (cellToBeDelete: ICell) {
 
-    const cellToBeDelete: IGetCellByPropsId = await getCellByProps_Id(cell_id);
+    // get all cells whit the same stem cell.
+    const cells: IGetCells = await getCellsByPropsIdStemCell(cellToBeDelete.idStemCell);
 
-    const positionCellToBeDelete:number = cellToBeDelete.cell_Request[0].position;
+    const positionCellToBeDelete:number = cellToBeDelete.position;
 
     const qteCell: number = cells.cells_Request.length*2;
 
-    for(let counter = positionCellToBeDelete; counter <= qteCell; counter+=2) {
+    // if cellToBeDelete is the main stem cell, the position of the sisters cells are not updated.
+    // else the position of the sisters cell are updated.
+    if(!cellToBeDelete.stemCell) {
+        for(let counter = positionCellToBeDelete; counter <= qteCell; counter+=2) {
 
-        const currentUpdateCell = await CellModel.find({position: counter, idStemCell: stemCell_id});
-      
-        await CellModel.findOneAndUpdate({_id: currentUpdateCell[0]._id}, {position: counter - 2})
-            .catch(error => console.log({
-                    request_type: `Update position of the cell ${currentUpdateCell} when cell is add`,
-                    error: true,
-                    message: error,
-                }))
+            const currentUpdateCell = await CellModel.find({position: counter, idStemCell: cellToBeDelete.idStemCell});
+          
+            await CellModel.findOneAndUpdate({_id: currentUpdateCell[0]._id}, {position: counter - 2})
+                .catch(error => console.log({
+                        request_type: `Update position of the cell ${currentUpdateCell} when cell is add`,
+                        error: true,
+                        message: error,
+                    }))
+        };
+    }
+
+    await CellModel.findByIdAndDelete(cellToBeDelete._id);
+
+    const parentTreeToBeDelete = await ParentsTreeOfTheCellModel.find({cellId: cellToBeDelete._id});
+    if(parentTreeToBeDelete.length > 0) {
+        await ParentsTreeOfTheCellModel.findByIdAndDelete(parentTreeToBeDelete[0]._id);
     };
 
-    await CellModel.findByIdAndDelete(cellToBeDelete.cell_Request[0]._id);
-
-    const parentTreeToBeDelete = await ParentsTreeOfTheCellModel.find({cellId:cellToBeDelete.cell_Request[0]._id});
-    await ParentsTreeOfTheCellModel.findByIdAndDelete(parentTreeToBeDelete[0]._id);
-    
-    const childrenIdList = await getAllIdOfChildCells(cell_id);
+    // delete all children of the cellToBeDeleted.
+    const childrenIdList = await getAllIdOfChildCells(cellToBeDelete._id);
     await deleteAllChildrenCellsOfTheCellDeleted(childrenIdList);
     await deleteAllParentsTreesOfTheCellDeleted(childrenIdList);
 }
+
+
+
